Turn on hold mode
I have no idea what this is, but it means we can skip the 5s sleep and
everything still works.  Each display takes about 0.1% of the battery
life.

Also removed unnecessary Paint_Clear call.  We're going to overwrite
the frame buffer with the bmp contents so there's no point in clearing
it out first.